Fixed compilation errors and warnings in unit test
Fixed an error due to a narrowing conversion in an initializer list in core_func_integer_bit_count.cpp, and various
printf format string issues that caused compiler warnings for me. std::clock_t isn't guaranteed to be any specific
integer type, nor is it even guaranteed to be integral (it can be floating point). So, the safest bet is to cast it to
double and use the %f specifier in printf format strings.