build: add pkg-config file
GLM used to ship a pkg-config file since version 0.9.8.0, added in
commit 0e018a5262b881e3b49b8a49ae56292ab9a7bb10, but it was removed
together with the install target in commit
5f352ecce21bb1ab37fa56fac0f383c779b351a3. While the install target was
re-added in commit 631faffab37bd14f0ac8cec74711d670538a2702, the .pc
file has been left behind.

This patch re-adds it, and, compared to the old one, this pkg-config
file also accounts for the cases where GLM is installed in an include
directory that is not a subdirectory of the install prefix (i.e. it does
the right thing™ when CMAKE_INSTALL_INCLUDEDIR is an absolute path).

To read why the JoinPaths module is needed, take a look at
<https://github.com/jtojnar/cmake-snips#concatenating-paths-when-building-pkg-config-files>